Refactor: 미사용 import 제거 및 FileHandler 모듈화 / 코드 수정
