+++ src/main/java/com/sideProject/PlanIT/domain/post/dto/NoticeDto.java
@@ -3,7 +3,6 @@ package com.sideProject.PlanIT.domain.post.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
