+++ src/main/java/com/sideProject/PlanIT/domain/post/repository/NoticeRepository.java
@@ -4,7 +4,6 @@ import com.sideProject.PlanIT.domain.post.entity.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
