--- src/main/java/com/sideProject/PlanIT/domain/post/service/FileService.java
@@ -1,12 +1,0 @@
-package com.sideProject.PlanIT.domain.post.service;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.File;
-import java.io.IOException;
-
-public interface FileService {
-    String saveFile(MultipartFile file);
-}
