+++ src/main/java/com/sideProject/PlanIT/domain/post/service/NoticeService.java
@@ -3,15 +3,7 @@ package com.sideProject.PlanIT.domain.post.service;
 
 import com.sideProject.PlanIT.domain.post.dto.NoticeDto;
 import com.sideProject.PlanIT.domain.post.entity.Notice;
-import com.sideProject.PlanIT.domain.post.repository.NoticeRepository;
-import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
