+++ src/main/java/com/sideProject/PlanIT/domain/post/service/NoticeServiceImpl.java
@@ -1,17 +1,13 @@
 package com.sideProject.PlanIT.domain.post.service;
 
 
+import com.sideProject.PlanIT.common.modules.FileHandler;
 import com.sideProject.PlanIT.domain.post.dto.NoticeDto;
 import com.sideProject.PlanIT.domain.post.entity.Notice;
 import com.sideProject.PlanIT.domain.post.repository.NoticeRepository;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -20,7 +16,7 @@ import java.util.stream.Collectors;
 public class NoticeServiceImpl implements NoticeService{
 
     private final NoticeRepository noticeRepository;
-    private final FileService fileService;
+    private final FileHandler fileHandler;
 
     @Override
     public Notice createNotice(NoticeDto.NoticeRequestDto noticeRequestDto) {
@@ -28,8 +24,8 @@ public class NoticeServiceImpl implements NoticeService{
                 .title(noticeRequestDto.getTitle())
                 .startAt(noticeRequestDto.getStartAt())
                 .endAt(noticeRequestDto.getEndAt())
-                .attachmentPath(fileService.saveFile(noticeRequestDto.getAttachment()))
-                .imagePath(fileService.saveFile(noticeRequestDto.getImage()))
+                .attachmentPath(fileHandler.saveFile(noticeRequestDto.getAttachment()))
+                .imagePath(fileHandler.saveFile(noticeRequestDto.getImage()))
                 .content(noticeRequestDto.getContent())
                 .build());
     }
@@ -37,8 +33,8 @@ public class NoticeServiceImpl implements NoticeService{
     @Override
     public Notice editNotice(Long notice_id, NoticeDto.NoticeRequestDto noticeRequestDto) {
         Notice noticeToEdit = noticeRepository.findById(notice_id).orElseThrow(() -> new IllegalArgumentException("no exist Id"));
-        fileService.saveFile(noticeRequestDto.getAttachment());
-        fileService.saveFile(noticeRequestDto.getImage());
+        fileHandler.saveFile(noticeRequestDto.getAttachment());
+        fileHandler.saveFile(noticeRequestDto.getImage());
         //todo: 기존 저장된 파일 제거?
 
         return noticeRepository.save(noticeToEdit.update(noticeRequestDto));
